+++ flux-tests/tests/pos/surface/test03.rs
@@ -1,11 +1,7 @@
 #![feature(register_tool)]
 #![register_tool(flux)]
 
-#[flux::sig(
-fn(x: &strg i32[@n]) -> i32
-ensures x: i32[n+1]
-)]
-pub fn inc(x: &mut i32) -> i32 {
+#[flux::sig(fn(x: &strg i32[@n]) ensures x: i32[n+1])]
+pub fn inc(x: &mut i32) {
     *x += 1;
-    0
 }
